fix the code for the connatix connector
