+++ .vscode/settings.json
@@ -41,5 +41,8 @@
     "editor.codeActionsOnSave": {
       "source.fixAll.stylelint": true
     }
-  }
+  },
+  "cSpell.words": [
+    "Connatix"
+  ]
 }
