+++ airbyte-integrations/connectors/source-connatix/integration_tests/configured_catalog.json
@@ -2,21 +2,12 @@
   "streams": [
     {
       "stream": {
-        "name": "customers",
+        "name": "connatix_ad_revenue_report_stream",
         "json_schema": {},
         "supported_sync_modes": ["full_refresh"]
       },
       "sync_mode": "full_refresh",
       "destination_sync_mode": "overwrite"
-    },
-    {
-      "stream": {
-        "name": "employees",
-        "json_schema": {},
-        "supported_sync_modes": ["full_refresh", "incremental"]
-      },
-      "sync_mode": "incremental",
-      "destination_sync_mode": "append"
     }
   ]
 }
