+++ airbyte-integrations/connectors/source-connatix/source_connatix/schemas/connatix_ad_revenue_report_stream.json
